Reorganize query engine tests
